requirements.txt
pyrogram
pymongo
python_dotenv